Validated the math of the startTime distribution
def d= Math.floor(Math.abs(Random.nextGaussian) * 1000.0 / 2.0).toLong
val (x, y) = (1 to 100000000).map(_ => d).partition(_ > 1000)
(y.size) / (x.size + y.size).toDouble // ~ 0.95 bingo

+++ loadtest/src/it/scala/narrative/LoadTest.scala
@@ -27,9 +27,8 @@ class LoadTest extends Simulation {
 
   // We want 95% of the traffic within the last hour (~2 STD)
   def startTime: Long = {
-    // TODO double check the math here
-    val oneHourInMillis = TimeUnit.HOURS.toMillis(1L)
-    delta - Math.floor(Math.abs(Random.nextGaussian) * oneHourInMillis / 2L).toLong
+    val oneHourInMillis = TimeUnit.HOURS.toMillis(1L).toDouble
+    delta - Math.floor(Math.abs(Random.nextGaussian) * oneHourInMillis / 2.0).toLong
   }
 
   // Someone looking at the latest analytics trends
